Validate a hub config file against a Infectious Disease Modeling Hubs schema

Validate a hub config file against a Infectious Disease Modeling Hubs schema

Usage

validate_config(
  hub_path = ".",
  config = c("tasks", "admin"),
  config_path = NULL,
  schema_version = "from_config",
  branch = "main",
  pretty_errors = TRUE
)

Arguments

hub_path

Path to a local hub directory.

config

Name of config file to validate. One of "tasks" or "admin".

config_path

Defaults to NULL which assumes all config files are in the hub-config directory in the root of hub directory. Argument config_path can be used to override default by providing a path to the config file to be validated.

schema_version

Character string specifying the json schema version to be used for validation. The default value "from_config" will use the version specified in the schema_version property of the config file. "latest" will use the latest version available in the Infectious Disease Modeling Hubs schemas repository. Alternartively, a specific version of a schema (e.g. "v0.0.1") can be specified.

branch

The branch of the Infectious Disease Modeling Hubs schemas repository from which to fetch schema. Defaults to "main".

pretty_errors

Whether to launch an informative pretty table of any validation errors in the Viewer panel. Defaults to TRUE.

Value

Returns the result of validation. If validation is successful, will return TRUE. If any validation errors are detected, returns FALSE with details of errors appended as a data.frame to an errors attribute. To access the errors table use attr(x, "errors") where x is the output of the function.

See also

Examples

# Valid config file
validate_config(
  hub_path = system.file(
    "testhubs/simple/",
    package = "hubUtils"
  ),
  config = "tasks"
)
#> ✔ Successfully validated config file /home/runner/work/_temp/Library/hubUtils/testhubs/simple/hub-config/tasks.json against schema <https://raw.githubusercontent.com/Infectious-Disease-Modeling-Hubs/schemas/main/v0.0.0.9/tasks-schema.json>
#> [1] TRUE
#> attr(,"config_path")
#> /home/runner/work/_temp/Library/hubUtils/testhubs/simple/hub-config/tasks.json
#> attr(,"schema_version")
#> [1] "v0.0.0.9"
#> attr(,"schema_url")
#> https://raw.githubusercontent.com/Infectious-Disease-Modeling-Hubs/schemas/main/v0.0.0.9/tasks-schema.json
# Config file with errors
config_path <- system.file("error-schema/tasks-errors.json",
  package = "hubUtils"
)
validate_config(config_path = config_path, config = "tasks", pretty_errors = FALSE)
#> Warning: Schema errors detected in config file
#> /home/runner/work/_temp/Library/hubUtils/error-schema/tasks-errors.json
#> validated against schema
#> <https://raw.githubusercontent.com/Infectious-Disease-Modeling-Hubs/schemas/main/v0.0.0.9/tasks-schema.json>
#> [1] FALSE
#> attr(,"errors")
#>                                                instancePath
#> 1          /rounds/0/model_tasks/0/task_ids/target/required
#> 2 /rounds/0/model_tasks/0/output_type/mean/type_id/optional
#> 3 /rounds/0/model_tasks/0/output_type/mean/type_id/required
#> 4          /rounds/0/model_tasks/0/output_type/mean/type_id
#> 5                                 /rounds/0/submissions_due
#> 6                           /rounds/0/submissions_due/start
#> 7                             /rounds/0/submissions_due/end
#> 8                                 /rounds/0/submissions_due
#>                                                                                                                                          schemaPath
#> 1                             #/properties/rounds/items/properties/model_tasks/items/properties/task_ids/properties/target/properties/required/type
#> 2 #/properties/rounds/items/properties/model_tasks/items/properties/output_type/properties/mean/properties/type_id/oneOf/0/properties/optional/type
#> 3 #/properties/rounds/items/properties/model_tasks/items/properties/output_type/properties/mean/properties/type_id/oneOf/1/properties/required/type
#> 4                            #/properties/rounds/items/properties/model_tasks/items/properties/output_type/properties/mean/properties/type_id/oneOf
#> 5                                                                             #/properties/rounds/items/properties/submissions_due/oneOf/0/required
#> 6                                                                #/properties/rounds/items/properties/submissions_due/oneOf/1/properties/start/type
#> 7                                                                  #/properties/rounds/items/properties/submissions_due/oneOf/1/properties/end/type
#> 8                                                                                        #/properties/rounds/items/properties/submissions_due/oneOf
#>    keyword params.type params.passingSchemas params.missingProperty
#> 1     type array, null                    NA                   <NA>
#> 2     type        null                    NA                   <NA>
#> 3     type        null                    NA                   <NA>
#> 4    oneOf        NULL                    NA                   <NA>
#> 5 required        NULL                    NA            relative_to
#> 6     type      string                    NA                   <NA>
#> 7     type      string                    NA                   <NA>
#> 8    oneOf        NULL                    NA                   <NA>
#>                                     message
#> 1                        must be array,null
#> 2                              must be null
#> 3                              must be null
#> 4    must match exactly one schema in oneOf
#> 5 must have required property 'relative_to'
#> 6                            must be string
#> 7                            must be string
#> 8    must match exactly one schema in oneOf
#>                                                                                                                                                                                                                                                                                                                                                      schema
#> 1                                                                                                                                                                                                                                                                                                                                               array, null
#> 2                                                                                                                                                                                                                                                                                                                                                      null
#> 3                                                                                                                                                                                                                                                                                                                                                      null
#> 4                                                                      When mean is required, property set to single element 'NA' array, When mean is optional, property set to null, array, null, NA, NA, 1, NA, When mean is required, property set to null, When mean is optional, property set to single element 'NA' array, null, array, NA, NA, NA, 1
#> 5                                                                                                                                                                                                                                                                                                                                   relative_to, start, end
#> 6                                                                                                                                                                                                                                                                                                                                                    string
#> 7                                                                                                                                                                                                                                                                                                                                                    string
#> 8 Name of task id variable in relation to which submission start and end dates are calculated., NA, string, NA, Difference in days between start and origin date., Submission start date., integer, string, NA, date, Difference in days between end and origin date., Submission end date., integer, string, NA, date, relative_to, start, end, start, end
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           parentSchema.description
#> 1                                                                                                                                                                                                                                                                                                                                         Array of target unique identifiers that must be present for submission to be valid. Can be null if no targets are required and all valid targets are specified in the optional property.
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      When mean is required, property set to null
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      When mean is optional, property set to null
#> 4 type_id is not meaningful for a mean output_type. The property is primarily used to determine whether mean is a required or optional output type through properties required and optional. If mean is a required output type, the required property must be an array containing the single string element 'NA' and the optional property must be set to null. If mean is an optional output type, the optional property must be an array containing the single string element 'NA' and the required property must be set to null
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             <NA>
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Submission start date.
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             Submission end date.
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                                                 Object defining the dates by which model forecasts must be submitted to the hub.
#>   parentSchema.type parentSchema.type
#> 1       array, null            string
#> 2              null              <NA>
#> 3              null              <NA>
#> 4            object              <NA>
#> 5              NULL              <NA>
#> 6            string              <NA>
#> 7            string              <NA>
#> 8            object              <NA>
#>                            parentSchema.examples
#> 1                                           NULL
#> 2                                           NULL
#> 3                                           NULL
#> 4                                         NA, NA
#> 5                                           NULL
#> 6                                           NULL
#> 7                                           NULL
#> 8 2022-06-07, -4, 2022-07-20, 2, NA, origin_date
#>                                                                                                                                                                                                                                                                                                                                          parentSchema.oneOf
#> 1                                                                                                                                                                                                                                                                                                                                                      NULL
#> 2                                                                                                                                                                                                                                                                                                                                                      NULL
#> 3                                                                                                                                                                                                                                                                                                                                                      NULL
#> 4                                                                      When mean is required, property set to single element 'NA' array, When mean is optional, property set to null, array, null, NA, NA, 1, NA, When mean is required, property set to null, When mean is optional, property set to single element 'NA' array, null, array, NA, NA, NA, 1
#> 5                                                                                                                                                                                                                                                                                                                                                      NULL
#> 6                                                                                                                                                                                                                                                                                                                                                      NULL
#> 7                                                                                                                                                                                                                                                                                                                                                      NULL
#> 8 Name of task id variable in relation to which submission start and end dates are calculated., NA, string, NA, Difference in days between start and origin date., Submission start date., integer, string, NA, date, Difference in days between end and origin date., Submission end date., integer, string, NA, date, relative_to, start, end, start, end
#>     parentSchema.required
#> 1                    NULL
#> 2                    NULL
#> 3                    NULL
#> 4      required, optional
#> 5 relative_to, start, end
#> 6                    NULL
#> 7                    NULL
#> 8              start, end
#>                                                parentSchema.properties.relative_to.description
#> 1                                                                                         <NA>
#> 2                                                                                         <NA>
#> 3                                                                                         <NA>
#> 4                                                                                         <NA>
#> 5 Name of task id variable in relation to which submission start and end dates are calculated.
#> 6                                                                                         <NA>
#> 7                                                                                         <NA>
#> 8                                                                                         <NA>
#>   parentSchema.properties.relative_to.type
#> 1                                     <NA>
#> 2                                     <NA>
#> 3                                     <NA>
#> 4                                     <NA>
#> 5                                   string
#> 6                                     <NA>
#> 7                                     <NA>
#> 8                                     <NA>
#>           parentSchema.properties.start.description
#> 1                                              <NA>
#> 2                                              <NA>
#> 3                                              <NA>
#> 4                                              <NA>
#> 5 Difference in days between start and origin date.
#> 6                                              <NA>
#> 7                                              <NA>
#> 8                                              <NA>
#>   parentSchema.properties.start.type
#> 1                               <NA>
#> 2                               <NA>
#> 3                               <NA>
#> 4                               <NA>
#> 5                            integer
#> 6                               <NA>
#> 7                               <NA>
#> 8                               <NA>
#>           parentSchema.properties.end.description
#> 1                                            <NA>
#> 2                                            <NA>
#> 3                                            <NA>
#> 4                                            <NA>
#> 5 Difference in days between end and origin date.
#> 6                                            <NA>
#> 7                                            <NA>
#> 8                                            <NA>
#>   parentSchema.properties.end.type parentSchema.format            data
#> 1                             <NA>                <NA> wk inc flu hosp
#> 2                             <NA>                <NA>              NA
#> 3                             <NA>                <NA>              NA
#> 4                             <NA>                <NA>          NA, NA
#> 5                          integer                <NA>           -6, 1
#> 6                             <NA>                date              -6
#> 7                             <NA>                date               1
#> 8                             <NA>                <NA>           -6, 1
#>                                                    dataPath
#> 1          /rounds/0/model_tasks/0/task_ids/target/required
#> 2 /rounds/0/model_tasks/0/output_type/mean/type_id/optional
#> 3 /rounds/0/model_tasks/0/output_type/mean/type_id/required
#> 4          /rounds/0/model_tasks/0/output_type/mean/type_id
#> 5                                 /rounds/0/submissions_due
#> 6                           /rounds/0/submissions_due/start
#> 7                             /rounds/0/submissions_due/end
#> 8                                 /rounds/0/submissions_due
#> attr(,"config_path")
#> [1] "/home/runner/work/_temp/Library/hubUtils/error-schema/tasks-errors.json"
#> attr(,"schema_version")
#> [1] "v0.0.0.9"
#> attr(,"schema_url")
#> https://raw.githubusercontent.com/Infectious-Disease-Modeling-Hubs/schemas/main/v0.0.0.9/tasks-schema.json

# Print pretty errors in help & pkgdown documentation. Not necessary when running
# interactively.
validate_config(config_path = config_path,
                config = "tasks", pretty_errors = FALSE) |>
                launch_pretty_errors_report() |>
                gt::as_raw_html()
#> Warning: Schema errors detected in config file
#> /home/runner/work/_temp/Library/hubUtils/error-schema/tasks-errors.json
#> validated against schema
#> <https://raw.githubusercontent.com/Infectious-Disease-Modeling-Hubs/schemas/main/v0.0.0.9/tasks-schema.json>
#> <div id="slvbyqbfpj" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   
#>   <table class="gt_table" style="font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif; display: table; border-collapse: collapse; margin-left: 2%; margin-right: 2%; color: #333333; font-size: 16px; font-weight: normal; font-style: normal; background-color: #FFFFFF; width: auto; border-top-style: solid; border-top-width: 2px; border-top-color: #A8A8A8; border-right-style: none; border-right-width: 2px; border-right-color: #D3D3D3; border-bottom-style: solid; border-bottom-width: 2px; border-bottom-color: #A8A8A8; border-left-style: none; border-left-width: 2px; border-left-color: #D3D3D3; table-layout: fixed;" bgcolor="#FFFFFF">
#>   <colgroup>
#>     <col>
#>     <col>
#>     <col>
#>     <col style="width:16.6666666666667%;">
#>     <col style="width:25%;">
#>     <col style="width:16.6666666666667%;">
#>   </colgroup>
#>   <thead class="gt_header">
#>     <tr>
#>       <td colspan="6" class="gt_heading gt_title gt_font_normal" style="background-color: #F0F3F5; text-align: center; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; color: #333333; font-size: 125%; padding-top: 4px; padding-bottom: 4px; padding-left: 5px; padding-right: 5px; border-bottom-color: #FFFFFF; border-bottom-width: 0; font-weight: normal;" bgcolor="#F0F3F5" align="center"><strong><code>hubUtils</code> config validation error report</strong></td>
#>     </tr>
#>     <tr>
#>       <td colspan="6" class="gt_heading gt_subtitle gt_font_normal gt_bottom_border" style="background-color: #F0F3F5; text-align: center; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; color: #333333; font-size: 85%; padding-top: 0; padding-bottom: 6px; padding-left: 5px; padding-right: 5px; border-top-color: #FFFFFF; border-top-width: 0; border-bottom-style: solid; border-bottom-width: 2px; border-bottom-color: #D3D3D3; font-weight: normal;" bgcolor="#F0F3F5" align="center">Report for file <strong><code>/home/runner/work/_temp/Library/hubUtils/error-schema/tasks-errors.json</code></strong> using
#> schema version <a href="https://raw.githubusercontent.com/Infectious-Disease-Modeling-Hubs/schemas/main/v0.0.0.9/tasks-schema.json"><strong>v0.0.0.9</strong></a></td>
#>     </tr>
#>   </thead>
#>   <thead class="gt_col_headings" style="border-top-style: solid; border-top-width: 2px; border-top-color: #D3D3D3; border-bottom-style: solid; border-bottom-width: 2px; border-bottom-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3;">
#>     <tr>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2" scope="colgroup" id="&lt;strong&gt;Error location&lt;/strong&gt;" style="color: #333333; background-color: #F0F3F5; font-size: 100%; font-weight: bold; text-transform: inherit; padding-top: 0; padding-bottom: 0; padding-right: 4px; text-align: center; padding-left: 0;" bgcolor="#F0F3F5" align="center">
#>         <span class="gt_column_spanner" style="border-bottom-style: solid; border-bottom-width: 2px; border-bottom-color: #D3D3D3; vertical-align: bottom; padding-top: 5px; padding-bottom: 5px; overflow-x: hidden; display: inline-block; width: 100%;"><strong>Error location</strong></span>
#>       </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="3" scope="colgroup" id="&lt;strong&gt;Schema details&lt;/strong&gt;" style="color: #333333; background-color: #F0F3F5; font-size: 100%; font-weight: bold; text-transform: inherit; padding-top: 0; padding-bottom: 0; padding-left: 4px; padding-right: 4px; text-align: center;" bgcolor="#F0F3F5" align="center">
#>         <span class="gt_column_spanner" style="border-bottom-style: solid; border-bottom-width: 2px; border-bottom-color: #D3D3D3; vertical-align: bottom; padding-top: 5px; padding-bottom: 5px; overflow-x: hidden; display: inline-block; width: 100%;"><strong>Schema details</strong></span>
#>       </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="1" scope="col" id="&lt;strong&gt;Config&lt;/strong&gt;" style="color: #333333; background-color: #F0F3F5; font-size: 100%; font-weight: bold; text-transform: inherit; padding-top: 0; padding-bottom: 0; padding-left: 4px; text-align: center; padding-right: 0;" bgcolor="#F0F3F5" align="center">
#>         <span class="gt_column_spanner" style="border-bottom-style: solid; border-bottom-width: 2px; border-bottom-color: #D3D3D3; vertical-align: bottom; padding-top: 5px; padding-bottom: 5px; overflow-x: hidden; display: inline-block; width: 100%;"><strong>Config</strong></span>
#>       </th>
#>     </tr>
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="instancePath" style="color: #333333; background-color: #F0F3F5; font-size: 100%; font-weight: bold; text-transform: inherit; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: bottom; padding-top: 5px; padding-bottom: 6px; padding-left: 5px; padding-right: 5px; overflow-x: hidden; text-align: left;" bgcolor="#F0F3F5" valign="bottom" align="left">instancePath</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="schemaPath" style="color: #333333; background-color: #F0F3F5; font-size: 100%; font-weight: bold; text-transform: inherit; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: bottom; padding-top: 5px; padding-bottom: 6px; padding-left: 5px; padding-right: 5px; overflow-x: hidden; text-align: left;" bgcolor="#F0F3F5" valign="bottom" align="left">schemaPath</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="keyword" style="color: #333333; background-color: #F0F3F5; font-size: 100%; font-weight: bold; text-transform: inherit; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: bottom; padding-top: 5px; padding-bottom: 6px; padding-left: 5px; padding-right: 5px; overflow-x: hidden; text-align: center;" bgcolor="#F0F3F5" valign="bottom" align="center">keyword</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="message" style="color: #333333; background-color: #F0F3F5; font-size: 100%; font-weight: bold; text-transform: inherit; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: bottom; padding-top: 5px; padding-bottom: 6px; padding-left: 5px; padding-right: 5px; overflow-x: hidden; text-align: center;" bgcolor="#F0F3F5" valign="bottom" align="center">message</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="schema" style="color: #333333; background-color: #F0F3F5; font-size: 100%; font-weight: bold; text-transform: inherit; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: bottom; padding-top: 5px; padding-bottom: 6px; padding-left: 5px; padding-right: 5px; overflow-x: hidden; text-align: left;" bgcolor="#F0F3F5" valign="bottom" align="left">schema</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="data" style="color: #333333; background-color: #F0F3F5; font-size: 100%; font-weight: bold; text-transform: inherit; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: bottom; padding-top: 5px; padding-bottom: 6px; padding-left: 5px; padding-right: 5px; overflow-x: hidden; text-align: center;" bgcolor="#F0F3F5" valign="bottom" align="center">data</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body" style="border-top-style: solid; border-top-width: 2px; border-top-color: #D3D3D3; border-bottom-style: solid; border-bottom-width: 2px; border-bottom-color: #D3D3D3;">
#>     <tr><td headers="instancePath" class="gt_row gt_left" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: left; white-space: pre;" valign="middle" align="left"><div class="gt_from_md"><p style="margin-top: 0; margin-bottom: 0;"><strong>rounds</strong>
#> └<strong>1</strong>
#> └─<strong>model_tasks</strong>
#> └──<strong>1</strong>
#> └───<strong>task_ids</strong>
#> └────<strong>target</strong>
#> └─────<strong>required</strong></p>
#> </div></td>
#> <td headers="schemaPath" class="gt_row gt_left" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: left; white-space: pre;" valign="middle" align="left"><div class="gt_from_md"><p style="margin-top: 0; margin-bottom: 0;">properties
#> └<strong>rounds</strong>
#> └─items
#> └──properties
#> └───<strong>model_tasks</strong>
#> └────items
#> └─────properties
#> └──────<strong>task_ids</strong>
#> └───────properties
#> └────────<strong>target</strong>
#> └─────────properties
#> └──────────<strong>required</strong>
#> └───────────<strong>type</strong></p>
#> </div></td>
#> <td headers="keyword" class="gt_row gt_center" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: center;" valign="middle" align="center">type</td>
#> <td headers="message" class="gt_row gt_center" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: center; background-color: #F9E3D6; font-weight: bold;" bgcolor="#F9E3D6" valign="middle" align="center">❌ must be array,null</td>
#> <td headers="schema" class="gt_row gt_left" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: left; white-space: pre-wrap;" valign="middle" align="left"><div class="gt_from_md"><p style="margin-top: 0; margin-bottom: 0;">array, null</p>
#> </div></td>
#> <td headers="data" class="gt_row gt_center" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: center; background-color: #F9E3D6; font-weight: bold;" bgcolor="#F9E3D6" valign="middle" align="center">wk inc flu hosp</td></tr>
#>     <tr><td headers="instancePath" class="gt_row gt_left" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: left; white-space: pre;" valign="middle" align="left"><div class="gt_from_md"><p style="margin-top: 0; margin-bottom: 0;"><strong>rounds</strong>
#> └<strong>1</strong>
#> └─<strong>model_tasks</strong>
#> └──<strong>1</strong>
#> └───<strong>output_type</strong>
#> └────<strong>mean</strong>
#> └─────<strong>type_id</strong></p>
#> </div></td>
#> <td headers="schemaPath" class="gt_row gt_left" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: left; white-space: pre;" valign="middle" align="left"><div class="gt_from_md"><p style="margin-top: 0; margin-bottom: 0;">properties
#> └<strong>rounds</strong>
#> └─items
#> └──properties
#> └───<strong>model_tasks</strong>
#> └────items
#> └─────properties
#> └──────<strong>output_type</strong>
#> └───────properties
#> └────────<strong>mean</strong>
#> └─────────properties
#> └──────────<strong>type_id</strong>
#> └───────────<strong>oneOf</strong></p>
#> </div></td>
#> <td headers="keyword" class="gt_row gt_center" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: center;" valign="middle" align="center">oneOf</td>
#> <td headers="message" class="gt_row gt_center" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: center; background-color: #F9E3D6; font-weight: bold;" bgcolor="#F9E3D6" valign="middle" align="center">❌ must match exactly one schema in oneOf</td>
#> <td headers="schema" class="gt_row gt_left" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: left; white-space: pre-wrap;" valign="middle" align="left"><div class="gt_from_md"><p style="margin-top: 0;"><strong>1</strong>
#> <strong>required-description:</strong> When mean is required, property set to single element 'NA' array
#> <strong>required-type:</strong> array
#> <strong>required-items-const:</strong>'NA'
#> <strong>required-items-maxItems:</strong> 1
#> <strong>optional-description:</strong> When mean is required, property set to null
#> <strong>optional-type:</strong> null</p>
#> <p style="margin-bottom: 0;"><strong>2</strong>
#> <strong>required-description:</strong> When mean is optional, property set to null
#> <strong>required-type:</strong> null
#> <strong>optional-description:</strong> When mean is optional, property set to single element 'NA' array
#> <strong>optional-type:</strong> array
#> <strong>optional-items-const:</strong>'NA'
#> <strong>optional-items-maxItems:</strong> 1</p>
#> </div></td>
#> <td headers="data" class="gt_row gt_center" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: center; background-color: #F9E3D6; font-weight: bold;" bgcolor="#F9E3D6" valign="middle" align="center">required: NA, optional: NA</td></tr>
#>     <tr><td headers="instancePath" class="gt_row gt_left" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: left; white-space: pre;" valign="middle" align="left"><div class="gt_from_md"><p style="margin-top: 0; margin-bottom: 0;"><strong>rounds</strong>
#> └<strong>1</strong>
#> └─<strong>submissions_due</strong></p>
#> </div></td>
#> <td headers="schemaPath" class="gt_row gt_left" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: left; white-space: pre;" valign="middle" align="left"><div class="gt_from_md"><p style="margin-top: 0; margin-bottom: 0;">properties
#> └<strong>rounds</strong>
#> └─items
#> └──properties
#> └───<strong>submissions_due</strong>
#> └────<strong>oneOf</strong></p>
#> </div></td>
#> <td headers="keyword" class="gt_row gt_center" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: center;" valign="middle" align="center">oneOf</td>
#> <td headers="message" class="gt_row gt_center" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: center; background-color: #F9E3D6; font-weight: bold;" bgcolor="#F9E3D6" valign="middle" align="center">❌ must match exactly one schema in oneOf</td>
#> <td headers="schema" class="gt_row gt_left" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: left; white-space: pre-wrap;" valign="middle" align="left"><div class="gt_from_md"><p style="margin-top: 0;"><strong>1</strong>
#> <strong>relative_to-description:</strong> Name of task id variable in relation to which submission start and end dates are calculated.
#> <strong>relative_to-type:</strong> string
#> <strong>start-description:</strong> Difference in days between start and origin date.
#> <strong>start-type:</strong> integer
#> <strong>start-format:</strong>'NA'
#> <strong>end-description:</strong> Difference in days between end and origin date.
#> <strong>end-type:</strong> integer
#> <strong>end-format:</strong>'NA'
#> <strong>required1:</strong> relative_to
#> <strong>required2:</strong> start
#> <strong>required3:</strong> end</p>
#> <p style="margin-bottom: 0;"><strong>2</strong>
#> <strong>relative_to-description:</strong>'NA'
#> <strong>relative_to-type:</strong>'NA'
#> <strong>start-description:</strong> Submission start date.
#> <strong>start-type:</strong> string
#> <strong>start-format:</strong> date
#> <strong>end-description:</strong> Submission end date.
#> <strong>end-type:</strong> string
#> <strong>end-format:</strong> date
#> <strong>required1:</strong> start
#> <strong>required2:</strong> end</p>
#> </div></td>
#> <td headers="data" class="gt_row gt_center" style="padding-top: 5px; padding-bottom: 5px; padding-left: 5px; padding-right: 5px; margin: 10px; border-top-style: solid; border-top-width: 1px; border-top-color: #D3D3D3; border-left-style: none; border-left-width: 1px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 1px; border-right-color: #D3D3D3; vertical-align: middle; overflow-x: hidden; text-align: center; background-color: #F9E3D6; font-weight: bold;" bgcolor="#F9E3D6" valign="middle" align="center">start: -6, end: 1</td></tr>
#>   </tbody>
#>   <tfoot class="gt_sourcenotes" style="color: #333333; background-color: #FFFFFF; border-bottom-style: none; border-bottom-width: 2px; border-bottom-color: #D3D3D3; border-left-style: none; border-left-width: 2px; border-left-color: #D3D3D3; border-right-style: none; border-right-width: 2px; border-right-color: #D3D3D3;" bgcolor="#FFFFFF">
#>     <tr>
#>       <td class="gt_sourcenote" colspan="6" style="font-size: 90%; padding-top: 4px; padding-bottom: 4px; padding-left: 5px; padding-right: 5px;">For more information, please consult the
#> <a href="https://hubdocs.readthedocs.io/en/latest/"><strong><code>hubDocs</code> documentation</strong>.</a></td>
#>     </tr>
#>   </tfoot>
#>   
#> </table>
#> </div>